requirements/local windows.txt
asgiref==3.6.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.16
django-ckeditor==5.9.0
django-import-export==3.0.2
django-js-asset==1.2.2
django-model-utils==4.0.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
PyYAML==6.0
sqlparse==0.3.0
tablib==3.3.0
typing-extensions==4.4.0
Unipath==1.1
xlrd==2.0.1
xlwt==1.3.0